feat(tooltips): enable tooltips for cpu and memory modules
Hi sejjy,

I've enabled the tooltips for the CPU and memory modules. This provides some extra information on hover, which I find quite useful.

It seems you already had support for the memory tooltip, but it was commented out. I've enabled it and also added it for the CPU module. It's a nice little addition that doesn't have any performance impact.

Let me know what you think